requirements.txt
bottle
python-magic
google-cloud-vision
pyexiftool
Pillow
pymongo[srv]
google-cloud-videointelligence
moviepy
backoff
watchdog
redis
numpy
tensorflow
sdnotify
